chore: convert few `jest-worker` tests to typescript
